Setting up getText
Trying to get the text from a txt file into a variable to be parsed
