+++ JavaScript Hide Text/getText.js
@@ -1,10 +1,36 @@
-import convert_text_to_binary from 'otherFunctions.js'
+/**
+	@brief	converts a string of ASCII characters to binary
+	@param	text	a string of ASCII encoded characters
+	@return	a string of binary bits
+*/
+function convert_text_to_binary(text) {
+    let output = ""
+    for(let i = 0; i < text.length; i++) {
+        output += text[i].charCodeAt(0) // .toString(2) + " "
+    }
+
+    output = parseInt(output)
+    console.log(output)
+    return output
+}
 
 /**
 	@brief	retrieves a text file from a HTML element
 	@return
 */
 function getText() {
-	let binary = convert_text_to_binary(test)
+	
+	let text = reader.readAsText(file);
+
+	let binary = convert_text_to_binary(text)
+	console.log(binary)
+	
+	return binary
 	
 }
+
+
+// TESTING AREA
+let test = "HELLO USER"
+
+let binary = convert_text_to_binary(test)
