+++ JavaScript Hide Text/index.1.html
@@ -1,0 +1,45 @@
+<html>
+<head> 
+    
+</head>
+<script>
+    var fileInput = $('#files');
+    var uploadButton = $('#upload');
+
+    uploadButton.on('click', function() {
+        if (!window.FileReader) {
+            alert('Your browser is not supported')
+        }
+        var input = fileInput.get(0);
+        
+        // Create a reader object
+        var reader = new FileReader();
+        if (input.files.length) {
+            var textFile = input.files[0];
+            reader.readAsText(textFile);
+            $(reader).on('load', processFile);
+        } else {
+            alert('Please upload a file before continuing')
+        } 
+    });
+
+    function processFile(e) {
+        var file = e.target.result,
+            results;
+        if (file && file.length) {
+            results = file.split("\n");
+            console.log(results)
+            $('#name').val(results[0]);
+            $('#age').val(results[1]);
+        }
+    }
+</script>
+
+<body>
+    <input type=file id=files />
+    <button id=upload>Upload</button>
+    <br/><br/>
+    <input type=text id=name /><br/>
+    <input type=text id=age />
+</body>
+</html>
