+++ JavaScript Hide Text/index.html
@@ -1,45 +1,45 @@
 <html>
-<head> 
+<div id="page-wrapper">
+    <script>
+    // source: https://codepen.io/matt-west/pen/KjEHg
+    var $injector = angular.injector(['q']);
     
-</head>
-<script>
-var fileInput = $('#files');
-var uploadButton = $('#upload');
+    window.onload = function() {
+		var fileInput = document.getElementById('fileInput');
+		var fileDisplayArea = document.getElementById('fileDisplayArea');
+        var deferred = $q.defer()
 
-uploadButton.on('click', function() {
-    if (!window.FileReader) {
-        alert('Your browser is not supported')
-    }
-    var input = fileInput.get(0);
-    
-    // Create a reader object
-    var reader = new FileReader();
-    if (input.files.length) {
-        var textFile = input.files[0];
-        reader.readAsText(textFile);
-        $(reader).on('load', processFile);
-    } else {
-        alert('Please upload a file before continuing')
-    } 
-});
+		fileInput.addEventListener('change', function(e) {
+			var file = fileInput.files[0];
+			var textType = /text.*/;
 
-function processFile(e) {
-    var file = e.target.result,
-        results;
-    if (file && file.length) {
-        results = file.split("\n");
-        console.log(results)
-        $('#name').val(results[0]);
-        $('#age').val(results[1]);
-    }
-}
-</script>
+			if (file.type.match(textType)) {
+                var reader = new FileReader();
 
-<body>
-    <input type=file id=files />
-    <button id=upload>Upload</button>
-    <br/><br/>
-    <input type=text id=name /><br/>
-    <input type=text id=age />
-</body>
+                reader.onload = function(e) {
+                    fileDisplayArea.innerText = reader.result;
+                    deferred.resolve(result);
+                }
+                
+            
+                reader.readAsText(file);
+                
+			} else {
+				fileDisplayArea.innerText = "File not supported!"
+            }
+            return deferred
+            
+		});
+    }    
+    </script>
+   
+   
+    <h1>Text File Reader</h1>
+    <div>
+        Select a text file: 
+        <input type="file" id="fileInput">
+    </div>
+    <pre id="fileDisplayArea"><pre>
+
+</div>
 </html>
