+++ JavaScript Hide Text/otherFunctions.js
@@ -1,20 +1,6 @@
 /**
-	@brief	converts a string of ASCII characters to binary
-	@param	text	a string of ASCII encoded characters
-	@return	a string of binary bits
-*/
-function convert_text_to_binary(text) {
-    let output = ""
-    for(let i = 0; i < text.length; i++) {
-        output += text[i].charCodeAt(0) // .toString(2) + " "
-    }
-
-    output = parseInt(output)
-    console.log(output)
-    return output
-}
-
-/**
+ * OLD FUNCTION. 
+ * Using strings instead the bits
 	@brief	checks the status of an individual bit of data within a byte
 	@param	letter	one byte of data (8 bits of value 0 or 1)
 	@param	index	the index of the bit we wish to check (0 = leftmost, 7 = rightmost)
@@ -42,6 +28,8 @@ function convert_binary_to_text(binary) {
 }
 
 
+// TESTING AREA
+
 let test = "HELLO USER"
 // let binary = convert_text_to_binary(letter)
 // console.log(binary) 
@@ -52,4 +40,4 @@ let test = "HELLO USER"
 
 let binary = convert_text_to_binary(test)
 
-// convert_binary_to_text(binary)
+convert_binary_to_text(binary)
