+++ JavaScript Hide Text/test.txt
@@ -1,0 +1,4 @@
+HELLO
+
+
+LKSDJFLKJSDF
